Convert player spawns to "control points" which produce energy and can be captured.

+++ src/data/entity.h
@@ -9,7 +9,7 @@ namespace VI
 
 typedef u8 Family;
 typedef u16 Revision;
-typedef u64 ComponentMask;
+typedef u32 ComponentMask;
 
 const Family MAX_FAMILIES = sizeof(ComponentMask) * 8;
 #define MAX_ENTITIES 2048
